requirements.txt
absl-py==1.4.0
anyascii==0.3.1
astunparse==1.6.3
blis==0.7.5
cachetools==5.3.0
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.6
contractions==0.1.73
cycler==0.11.0
cymem==2.0.6
Cython==0.29.33
dm-tree==0.1.8
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.1.0/en_core_web_lg-3.1.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
etils==1.0.0
flatbuffers==23.3.3
fonttools==4.28.5
gast==0.4.0
gensim==4.1.2
gin-config==0.5.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.51.3
h5py==3.8.0
httplib2==0.21.0
idna==3.3
immutabledict==2.2.3
importlib-resources==5.12.0
jax==0.4.5
Jinja2==3.0.3
joblib==1.1.0
kaggle==1.5.13
keras-nightly==2.13.0.dev2023030908
kiwisolver==1.3.2
langcodes==3.3.0
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.1
murmurhash==1.0.6
negspacy==1.0.1
nltk==3.6.6
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pathy==0.6.1
Pillow==8.4.0
portalocker==2.7.0
preshed==3.0.6
promise==2.3
protobuf==3.20.3
psutil==5.9.4
py-cpuinfo==9.0.0
pyahocorasick==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pydantic==1.8.2
PyMySQL==1.0.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2021.3
PyYAML==5.4.1
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.3.1
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.97
seqeval==1.2.2
six==1.16.0
smart-open==5.2.1
spacy==3.1.4
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
tabulate==0.9.0
tb-nightly==2.13.0a20230309
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.19.0
tensorflow-datasets==4.8.3
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.12.0
tensorflow-model-optimization==0.7.3
tensorflow-text-nightly==2.13.0.dev20230306
tensorrt==8.5.3.1
termcolor==2.2.0
text-unidecode==1.3
textblob==0.17.1
textsearch==0.0.24
tf-estimator-nightly==2.13.0.dev2023030909
tf-hub-nightly==0.13.0.dev202303060822
tf-models-nightly==2.11.0.dev20230306
tf-nightly==2.13.0.dev20230309
tf-slim==1.1.0
thinc==8.0.13
threadpoolctl==3.0.0
toml==0.10.2
torch==1.13.1
tqdm==4.62.3
typeguard==2.13.3
typer==0.4.0
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wasabi==0.9.0
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
